Unify APIs again
